refactor: MemberList 분리
